+++ Prism/Plugins/Apps/Houdini/Scripts/Prism_Houdini_Renderer_Redshift.py
@@ -11,23 +11,24 @@
 ####################################################
 #
 #
-# Copyright (C) 2016-2020 Richard Frangenberg
+# Copyright (C) 2016-2023 Richard Frangenberg
+# Copyright (C) 2023 Prism Software GmbH
 #
-# Licensed under GNU GPL-3.0-or-later
+# Licensed under GNU LGPL-3.0-or-later
 #
 # This file is part of Prism.
 #
 # Prism is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
+# it under the terms of the GNU Lesser General Public License as published by
 # the Free Software Foundation, either version 3 of the License, or
 # (at your option) any later version.
 #
 # Prism is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
+# GNU Lesser General Public License for more details.
 #
-# You should have received a copy of the GNU General Public License
+# You should have received a copy of the GNU Lesser General Public License
 # along with Prism.  If not, see <https://www.gnu.org/licenses/>.
 
 
@@ -37,13 +38,9 @@ import glob
 
 import hou
 
-try:
-    from PySide2.QtCore import *
-    from PySide2.QtGui import *
-    from PySide2.QtWidgets import *
-except:
-    from PySide.QtCore import *
-    from PySide.QtGui import *
+from qtpy.QtCore import *
+from qtpy.QtGui import *
+from qtpy.QtWidgets import *
 
 
 label = "Redshift"
@@ -56,6 +53,11 @@ def isActive():
 
 def getCam(node):
     return hou.node(node.parm("RS_renderCamera").eval())
+
+
+def getFormatFromNode(node):
+    fmt = node.parm("RS_outputFileFormat").evalAsString()
+    return fmt
 
 
 def createROP(origin):
@@ -166,28 +168,45 @@ def executeAOVs(origin, outputName):
     ):
         renderRS = True
 
-        rsOutput = os.path.join(os.path.dirname(outputName), "_rs", os.path.basename(outputName))
+        rsOutput = os.path.join(
+            os.path.dirname(outputName), "_rs", os.path.basename(outputName)
+        )
         rsOutput = os.path.splitext(rsOutput)[0] + ".rs"
-        if not origin.core.appPlugin.setNodeParm(origin.node, "RS_archive_file", val=rsOutput):
-            return [origin.state.text(0) + ": error - could not set archive filename. Publish canceled"]
+        parmPath = origin.core.appPlugin.getPathRelativeToProject(rsOutput) if origin.core.appPlugin.getUseRelativePath() else rsOutput
+        if not origin.core.appPlugin.setNodeParm(
+            origin.node, "RS_archive_file", val=parmPath
+        ):
+            return [
+                origin.state.text(0)
+                + ": error - could not set archive filename. Publish canceled"
+            ]
 
         os.makedirs(os.path.dirname(rsOutput))
 
     else:
         renderRS = False
 
-    if not origin.core.appPlugin.setNodeParm(origin.node, "RS_archive_enable", val=renderRS):
-        return [origin.state.text(0) + ": error - could not set archive enabled. Publish canceled"]
+    if not origin.core.appPlugin.setNodeParm(
+        origin.node, "RS_archive_enable", val=renderRS
+    ):
+        return [
+            origin.state.text(0)
+            + ": error - could not set archive enabled. Publish canceled"
+        ]
 
     if not origin.core.appPlugin.setNodeParm(origin.node, "RS_outputEnable", val=True):
         return [origin.state.text(0) + ": error - Publish canceled"]
 
+    parmPath = origin.core.appPlugin.getPathRelativeToProject(outputName) if origin.core.appPlugin.getUseRelativePath() else outputName
     if not origin.core.appPlugin.setNodeParm(
-        origin.node, "RS_outputFileNamePrefix", val=outputName
+        origin.node, "RS_outputFileNamePrefix", val=parmPath
     ):
         return [origin.state.text(0) + ": error - Publish canceled"]
 
-    origin.core.appPlugin.setNodeParm(origin.node, "RS_outputDisableSuffixes", val=1, severity="debug")
+    if origin.node.parm("RS_outputMultilayerMode").eval() != "2":
+        origin.core.appPlugin.setNodeParm(
+            origin.node, "RS_outputDisableSuffixes", val=1, severity="debug"
+        )
 
     for parm in origin.node.parms():
         if "RS_aovCustomPrefix" in parm.name():
@@ -204,7 +223,7 @@ def executeAOVs(origin, outputName):
             ):
                 return [origin.state.text(0) + ": error - Publish canceled"]
 
-            if origin.node.parm("RS_outputMultilayerMode").eval() == 1:
+            if origin.node.parm("RS_outputMultilayerMode").eval() == "2":
                 outPut = ""
 
             if not origin.core.appPlugin.setNodeParm(
@@ -216,10 +235,16 @@ def executeAOVs(origin, outputName):
     if ext in [".exr", ".png", ".jpg"]:
         formatVal = ext
     else:
-        return [origin.state.text(0) + ": error - invalid image format. Publish canceled"]
+        return [
+            origin.state.text(0) + ": error - invalid image format. Publish canceled"
+        ]
 
-    if not origin.core.appPlugin.setNodeParm(origin.node, "RS_outputFileFormat", val=formatVal):
-        return [origin.state.text(0) + ": error - could not set format. Publish canceled"]
+    if not origin.core.appPlugin.setNodeParm(
+        origin.node, "RS_outputFileFormat", val=formatVal
+    ):
+        return [
+            origin.state.text(0) + ": error - could not set format. Publish canceled"
+        ]
 
     return True
 
@@ -256,13 +281,13 @@ def waitForRenderCompleted(node):
     outputPath = node.parm("RS_outputFileNamePrefix").eval()
     outputDir = os.path.dirname(outputPath)
     globPath = outputDir + "/*.lock"
-    waitTime = 0
-    maxWaitTime = 10
+    # waitTime = 0
+    # maxWaitTime = 10
     while glob.glob(globPath):
         time.sleep(1)
-        waitTime += 1
-        if waitTime >= maxWaitTime:
-            return False
+        # waitTime += 1
+        # if waitTime >= maxWaitTime:
+            # return False
 
     return True
 
@@ -284,8 +309,11 @@ delDir = os.path.dirname(rsOutput)
 if os.path.basename(delDir) != "_rs":
     raise RuntimeError("invalid rs directory: %s" % (delDir))
 
-shutil.rmtree(delDir)
-print("task completed successfully")
+if os.path.exists(delDir):
+    shutil.rmtree(delDir)
+    print("task completed successfully")
+else:
+    print("directory doesn't exist")
 
 """
     return script
